Terminada

+++ minmax/src/function.h
@@ -1,9 +1,10 @@
 #ifndef FUNCTION_H
 #define FUNCTION_H
 
+#include <cstddef>
 #include <utility>
 using std::pair;
 
-std::pair<int, int> min_max( int V[], int n );
+std::pair<int, int> min_max( int V[], size_t n );
 
 #endif
